Rename location_features.py to features.py
